process-data/requirements.txt
pandas
openpyxl
boto3
io
tempfile
regex
hdx-python-api==6.1.4

